--- src/AdsConsentStatus.ts
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2016-present Invertase Limited & Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this library except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-/**
- * AdsConsentStatus enum.
- */
-export enum AdsConsentStatus {
-  /**
-   * Unknown consent status, AdsConsent.requestInfoUpdate needs to be called to update it.
-   */
-  UNKNOWN = 'UNKNOWN',
-
-  /**
-   * User consent required but not yet obtained.
-   */
-  REQUIRED = 'REQUIRED',
-
-  /**
-   * User consent not required.
-   */
-  NOT_REQUIRED = 'NOT_REQUIRED',
-
-  /**
-   * User consent already obtained.
-   */
-  OBTAINED = 'OBTAINED',
-}
